fix(agent): gate PSU Event Hub import and make it non-fatal
The PowerShell Universal Event Hub compatibility import ran on every agent
startup. It could enable an experimental feature solely from the presence of a
third-party PSU config file or `PSU_` environment variable, and could abort
startup entirely if such a file or variable was malformed.

The import is now gated at the call site behind both the `enable_unstable`
debug flag and an already-enabled PsuEventHub configuration. It never enables
the feature on its own: a stray PSU config or `PSU_` variable can no longer
bring the feature into existence. When the feature is explicitly enabled but
defines no connections of its own, the legacy PSU configuration is used to fill
them in; any explicit configuration always wins. The import is also best-effort:
any failure is logged and ignored instead of blocking startup.

The foreign-format PSU parsing is additionally moved out of the agent config
module into a dedicated `psu_event_hub::compat` module to keep legacy
migration code separate from the native agent configuration.
